backend/requirements.txt
fastapi
uvicorn[standard]
openai
pydantic
python-dotenv
pyyaml
requests
httpx
google-adk
google-genai
litellm